+++ packages/modal/tests/collection.test.js
@@ -1,5 +1,5 @@
 import "@testing-library/jest-dom/vitest";
-import Modal from "../index";
+import { ModalCollection } from "../index";
 
 const markup = `
   <div id="modal-1" class="modal">
@@ -11,19 +11,19 @@ const markup = `
 `;
 
 describe("register() & entry.deregister()", () => {
-  let modal, entry1, entry2, el1, el2;
+  let modals, entry1, entry2, el1, el2;
 
   beforeAll(() => {
     document.body.innerHTML = markup;
-    modal = new Modal({ transition: false });
+    modals = new ModalCollection({ transition: false });
     el1 = document.querySelector("#modal-1");
     el2 = document.querySelector("#modal-2");
   });
 
   it("should create collection entry when register is called with modal ID", async () => {
-    expect(modal.collection.length).toBe(0);
-    entry1 = await modal.register(await modal.createEntry("modal-1"));
-    expect(modal.collection.length).toBe(1);
+    expect(modals.collection.length).toBe(0);
+    entry1 = await modals.register(await modals.createEntry("modal-1"));
+    expect(modals.collection.length).toBe(1);
     expect(entry1.el).toBe(el1);
   });
 
@@ -34,42 +34,42 @@ describe("register() & entry.deregister()", () => {
   });
 
   it("should register modal without tabindex if setTabindex is disabled", async () => {
-    modal.config.setTabindex = false;
-    entry2 = await modal.register(await modal.createEntry("modal-2"));
+    modals.config.setTabindex = false;
+    entry2 = await modals.register(await modals.createEntry("modal-2"));
     expect(entry1.dialog.getAttribute("tabindex")).toBe("-1");
     expect(entry2.el).toBe(el2);
   });
 
   it("should not be able to register the same modal multiple times", async () => {
-    expect(modal.collection.length).toBe(2);
-    expect(modal.collection[0].id).toBe("modal-1");
-    expect(modal.collection[1].id).toBe("modal-2");
-    await modal.register(await modal.createEntry("modal-1"));
-    expect(modal.collection.length).toBe(2);
-    expect(modal.collection[0].id).toBe("modal-1");
-    expect(modal.collection[1].id).toBe("modal-2");
+    expect(modals.collection.length).toBe(2);
+    expect(modals.collection[0].id).toBe("modal-1");
+    expect(modals.collection[1].id).toBe("modal-2");
+    await modals.register(await modals.createEntry("modal-1"));
+    expect(modals.collection.length).toBe(2);
+    expect(modals.collection[0].id).toBe("modal-1");
+    expect(modals.collection[1].id).toBe("modal-2");
   });
 
   it("should remove entry from collection when entry method deregister is run", async () => {
-    entry1 = modal.get("modal-1");
-    entry2 = modal.get("modal-2");
+    entry1 = modals.get("modal-1");
+    entry2 = modals.get("modal-2");
     await entry1.deregister();
     await entry2.deregister();
-    expect(modal.collection.length).toBe(0);
+    expect(modals.collection.length).toBe(0);
   });
 });
 
 describe("entry.open() & entry.close() & entry.replace()", () => {
-  let modal;
+  let modals;
 
   beforeAll(async () => {
     document.body.innerHTML = markup;
-    modal = new Modal({ transition: false });
-    await modal.mount();
+    modals = new ModalCollection({ transition: false });
+    await modals.mount();
   });
 
   it("should open modal with transitions disabled", async () => {
-    const entry = modal.get("modal-1");
+    const entry = modals.get("modal-1");
     expect(entry.state).toBe("closed");
     expect(entry.el).toHaveClass("is-closed");
     await entry.open();
@@ -78,7 +78,7 @@ describe("entry.open() & entry.close() & entry.replace()", () => {
   });
 
   it("should close modal with transitions disabled", async () => {
-    const entry = modal.get("modal-1");
+    const entry = modals.get("modal-1");
     expect(entry.state).toBe("opened");
     expect(entry.el).toHaveClass("is-opened");
     await entry.close();
@@ -87,8 +87,8 @@ describe("entry.open() & entry.close() & entry.replace()", () => {
   });
 
   it("should replace modal with transitions disabled", async () => {
-    const entry1 = modal.get("modal-1");
-    const entry2 = modal.get("modal-2");
+    const entry1 = modals.get("modal-1");
+    const entry2 = modals.get("modal-2");
 
     await entry1.open();
     expect(entry1.state).toBe("opened");