+++ packages/modal/tests/stack.test.js
@@ -1,8 +1,8 @@
 import "@testing-library/jest-dom/vitest";
 import "./mocks/getComputedStyle.mock";
-import Modal from "../index";
+import { ModalCollection } from "../index";
 
-let modal, modal1, modal2, modal3, btn1, btn2, btn3, btn4, btnCloseAll;
+let modals, modal1, modal2, modal3, btn1, btn2, btn3, btn4, btnCloseAll;
 
 document.body.innerHTML = `
   <button data-modal-open="modal-1">...</button>
@@ -25,12 +25,12 @@ document.body.innerHTML = `
 `;
 
 beforeAll(async () => {
-  modal = new Modal({ transition: false });
-  await modal.mount();
+  modals = new ModalCollection({ transition: false });
+  await modals.mount();
 
-  modal1 = modal.get("modal-1");
-  modal2 = modal.get("modal-2");
-  modal3 = modal.get("modal-3");
+  modal1 = modals.get("modal-1");
+  modal2 = modals.get("modal-2");
+  modal3 = modals.get("modal-3");
 
   btn1 = document.querySelector('[data-modal-open="modal-1"]');
   btn2 = document.querySelector('[data-modal-open="modal-2"]');
@@ -40,14 +40,14 @@ beforeAll(async () => {
 });
 
 test("should allow opening multiple modals at once", async () => {
-  expect(modal.collection.length).toBe(3);
-  expect(modal.stack.length).toBe(0);
+  expect(modals.collection.length).toBe(3);
+  expect(modals.stack.length).toBe(0);
 
   btn1.click();
   btn2.click();
   btn3.click();
 
-  expect(modal.stack.length).toBe(3);
+  expect(modals.stack.length).toBe(3);
   expect(modal1.state).toBe("opened");
   expect(modal2.state).toBe("opened");
   expect(modal3.state).toBe("opened");
@@ -78,16 +78,16 @@ test("should have set focus to the correct modal dialog when stack order is chan
 test("should close the currently opened modal at the top of the stack", async () => {
   expect(modal2.state).toBe("opened");
 
-  modal.close();
+  modals.close();
 
   expect(modal2.state).toBe("closed");
   expect(document.activeElement).toBe(modal3.dialog);
 });
 
 test("should update the stack array and z-index of remaining active modals", () => {
-  expect(modal.stack.length).toBe(2);
-  expect(modal.stack.get(0)).toBe(modal1);
-  expect(modal.stack.get(1)).toBe(modal3);
+  expect(modals.stack.length).toBe(2);
+  expect(modals.stack.get(0)).toBe(modal1);
+  expect(modals.stack.get(1)).toBe(modal3);
 
   expect(modal1.el.style.zIndex).toBe("1001");
   expect(modal2.el.style.zIndex).toBe("");
@@ -97,13 +97,13 @@ test("should update the stack array and z-index of remaining active modals", () 
 test("should close the currently opened modal and update stack array and z-index styles", async () => {
   expect(modal3.state).toBe("opened");
 
-  modal.close();
+  modals.close();
 
   expect(modal3.state).toBe("closed");
   expect(document.activeElement).toBe(modal1.dialog);
 
-  expect(modal.stack.length).toBe(1);
-  expect(modal.stack.get(0)).toBe(modal1);
+  expect(modals.stack.length).toBe(1);
+  expect(modals.stack.get(0)).toBe(modal1);
 
   expect(modal1.el.style.zIndex).toBe("1001");
   expect(modal2.el.style.zIndex).toBe("");
@@ -113,12 +113,12 @@ test("should close the currently opened modal and update stack array and z-index
 test("should focus root trigger when the last modal in stack is closed", async () => {
   expect(modal1.state).toBe("opened");
 
-  modal.close();
+  modals.close();
 
   expect(modal1.state).toBe("closed");
   expect(document.activeElement).toBe(btn1);
 
-  expect(modal.stack.length).toBe(0);
+  expect(modals.stack.length).toBe(0);
 });
 
 test("should close all open modals when close button with value of * is set", async () => {
@@ -126,14 +126,14 @@ test("should close all open modals when close button with value of * is set", as
   btn2.click();
   btn3.click();
 
-  expect(modal.stack.length).toBe(3);
+  expect(modals.stack.length).toBe(3);
   expect(modal1.state).toBe("opened");
   expect(modal2.state).toBe("opened");
   expect(modal3.state).toBe("opened");
 
   btnCloseAll.click();
 
-  expect(modal.stack.length).toBe(0);
+  expect(modals.stack.length).toBe(0);
   expect(modal1.state).toBe("closed");
   expect(modal2.state).toBe("closed");
   expect(modal3.state).toBe("closed");
@@ -143,14 +143,14 @@ test("should properly replace all modal modals with the trigger modal", async ()
   btn1.click();
   btn2.click();
 
-  expect(modal.stack.length).toBe(2);
+  expect(modals.stack.length).toBe(2);
   expect(modal1.state).toBe("opened");
   expect(modal2.state).toBe("opened");
   expect(modal3.state).toBe("closed");
 
-  await modal.replace("modal-3");
+  await modals.replace("modal-3");
 
-  expect(modal.stack.length).toBe(1);
+  expect(modals.stack.length).toBe(1);
   expect(modal1.state).toBe("closed");
   expect(modal2.state).toBe("closed");
   expect(modal3.state).toBe("opened");