+++ packages/popover/tests/helpers.test.js
@@ -1,5 +1,5 @@
 import "@testing-library/jest-dom/vitest";
-import Popover from "../index";
+import { PopoverCollection } from "../index";
 import { cssConfig } from "@vrembem/core";
 import {
   applyPositionStyle,
@@ -10,7 +10,7 @@ import {
 } from "../src/js/helpers";
 import { expect } from "vitest";
 
-let popover;
+let popovers;
 
 const markup = `
   <button aria-controls="pop-1">...</button>
@@ -48,10 +48,10 @@ beforeAll(() => {
 });
 
 afterEach(() => {
-  if (popover && "unmount" in popover) {
-    popover.unmount();
+  if (popovers && "unmount" in popovers) {
+    popovers.unmount();
   }
-  popover = null;
+  popovers = null;
   document.body.innerHTML = null;
 });
 
@@ -126,40 +126,40 @@ describe("getPadding()", () => {
 describe("getPopoverID()", () => {
   it("should return the popover id using a popover", () => {
     document.body.innerHTML = markup;
-    popover = new Popover();
+    popovers = new PopoverCollection();
     const el = document.querySelector(".popover");
-    const result = getPopoverID.call(popover, el);
+    const result = getPopoverID.call(popovers, el);
     expect(result).toBe("pop-1");
   });
 
   it("should return the popover id using a popover trigger", () => {
     document.body.innerHTML = markup;
-    popover = new Popover();
+    popovers = new PopoverCollection();
     const el = document.querySelector('[aria-controls="pop-2"]');
-    const result = getPopoverID.call(popover, el);
+    const result = getPopoverID.call(popovers, el);
     expect(result).toBe("pop-2");
   });
 
   it("should return the popover id using a popover tooltip trigger", () => {
     document.body.innerHTML = markup;
-    popover = new Popover();
+    popovers = new PopoverCollection();
     const el = document.querySelector('[aria-describedby="pop-3"]');
-    const result = getPopoverID.call(popover, el);
+    const result = getPopoverID.call(popovers, el);
     expect(result).toBe("pop-3");
   });
 
   it("should return null if html element does not have the correct attributes", () => {
     document.body.innerHTML = markup;
-    popover = new Popover();
+    popovers = new PopoverCollection();
     const trigger = document.querySelector("#missing-attribute");
-    const result = getPopoverID.call(popover, trigger);
+    const result = getPopoverID.call(popovers, trigger);
     expect(result).toBe(null);
   });
 
   it("should return null if passed object does not resolve an ID", () => {
     document.body.innerHTML = markup;
-    popover = new Popover();
-    const result = getPopoverID.call(popover, true);
+    popovers = new PopoverCollection();
+    const result = getPopoverID.call(popovers, true);
     expect(result).toBe(null);
   });
 });
@@ -169,16 +169,16 @@ describe("getPopoverElements()", () => {
     document.body.innerHTML = markup;
     const trigger = document.querySelector('[aria-controls="pop-1"]');
     const target = document.querySelector("#pop-1");
-    popover = new Popover();
-    const result = getPopoverElements.call(popover, "pop-1");
+    popovers = new PopoverCollection();
+    const result = getPopoverElements.call(popovers, "pop-1");
     expect(result.popover).toBe(target);
     expect(result.trigger).toBe(trigger);
   });
 
   it("should throw error if no popover elements are found using an ID", () => {
     document.body.innerHTML = markup;
-    popover = new Popover();
-    expect(() => getPopoverElements.call(popover, "pop-4")).toThrow(
+    popovers = new PopoverCollection();
+    expect(() => getPopoverElements.call(popovers, "pop-4")).toThrow(
       'No popover elements found using the ID: "pop-4".'
     );
   });
@@ -186,8 +186,8 @@ describe("getPopoverElements()", () => {
   it("should throw error if no popover is found using a trigger element", () => {
     document.body.innerHTML = markup;
     const trigger = document.querySelector('[aria-controls="asdf"]');
-    popover = new Popover();
-    expect(() => getPopoverElements.call(popover, trigger)).toThrow(
+    popovers = new PopoverCollection();
+    expect(() => getPopoverElements.call(popovers, trigger)).toThrow(
       'No popover associated with the provided popover trigger: "asdf".'
     );
   });
@@ -195,8 +195,8 @@ describe("getPopoverElements()", () => {
   it("should throw error if no popover trigger is found using a popover", () => {
     document.body.innerHTML = markup;
     const target = document.querySelector("#fdsa");
-    popover = new Popover();
-    expect(() => getPopoverElements.call(popover, target)).toThrow(
+    popovers = new PopoverCollection();
+    expect(() => getPopoverElements.call(popovers, target)).toThrow(
       'No popover trigger associated with the provided popover: "fdsa".'
     );
   });
@@ -204,8 +204,8 @@ describe("getPopoverElements()", () => {
   it("should throw error if unable to resolve a popover ID with provided query", () => {
     document.body.innerHTML = markup;
     const trigger = document.querySelector("#missing-attribute");
-    popover = new Popover();
-    expect(() => getPopoverElements.call(popover, trigger)).toThrow(
+    popovers = new PopoverCollection();
+    expect(() => getPopoverElements.call(popovers, trigger)).toThrow(
       "Could not resolve the popover ID."
     );
   });
@@ -214,13 +214,13 @@ describe("getPopoverElements()", () => {
 describe("getDelay()", () => {
   it("should return the appropriate delay based on the provided index", async () => {
     document.body.innerHTML = markup;
-    popover = new Popover({
+    popovers = new PopoverCollection({
       toggleDelay: 200
     });
-    const entry1 = await popover.register(await popover.createEntry("pop-1"));
+    const entry1 = await popovers.register(await popovers.createEntry("pop-1"));
     expect(getDelay(entry1, 0)).toBe(200);
 
-    const entry2 = popover.get("pop-1");
+    const entry2 = popovers.get("pop-1");
     entry2.config.set({
       toggleDelay: [300, 600]
     });
@@ -230,22 +230,22 @@ describe("getDelay()", () => {
 
   it("should create an array if the provided delay is a string", async () => {
     document.body.innerHTML = customPropertyMarkup;
-    popover = new Popover({
+    popovers = new PopoverCollection({
       plugins: [cssConfig()]
     });
-    await popover.mount();
-    expect(getDelay(popover.get("asdf"), 0)).toBe(200);
-    expect(getDelay(popover.get("asdf"), 1)).toBe(400);
-    expect(getDelay(popover.get("fdsa"), 0)).toBe(400);
-    expect(getDelay(popover.get("fdsa"), 1)).toBe(800);
+    await popovers.mount();
+    expect(getDelay(popovers.get("asdf"), 0)).toBe(200);
+    expect(getDelay(popovers.get("asdf"), 1)).toBe(400);
+    expect(getDelay(popovers.get("fdsa"), 0)).toBe(400);
+    expect(getDelay(popovers.get("fdsa"), 1)).toBe(800);
   });
 
   it("should throw an error if the provided delay is not a number", async () => {
     document.body.innerHTML = markup;
-    popover = new Popover({
+    popovers = new PopoverCollection({
       toggleDelay: "asdf"
     });
-    const entry1 = await popover.register(await popover.createEntry("pop-1"));
+    const entry1 = await popovers.register(await popovers.createEntry("pop-1"));
     expect(() => getDelay(entry1, 0)).toThrow(
       'Provided delay value is not a number: "asdf"'
     );